fix: Make planner more conservative when overtaking
Planner need another state to overtake safely.
Namely it should first match the speed of the vehicle behind (if its
possible with a vehicle infront)
and then pass.
The other problem is that due to design, acceleration is always zero at
plan duration endpoints. This means that the car loses all the momentum
it has. This is something that need to be explored.